move schemaless flavor to dashboards model
